Removed the abstract qualifier for some BaseArFragment functions to lightweight the override.
